Update recommended mapping with valid inline comments.
The inline comments in the example break the commands. The `"` in the line is interpreted as a literal key-press.  Vimrc allows using the `|` to end the map command. This allows the changes to be directly copied into a user's vimrc.

From `:help map-comments:`
> It is not possible to put a comment after these commands, because the '"' character is considered to be part of the {lhs} or {rhs}.

Stack Overflow explaining more. [link](https://stackoverflow.com/a/24717020)